tests/wpt/tests/fetch/metadata/generated/css-font-face.https.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/css-font-face.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for CSS font-face</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  let count = 0;

  function induceRequest(t, url) {
    const id = `el-${count += 1}`;
    const style = document.createElement('style');
    style.appendChild(document.createTextNode(`
      @font-face {
        font-family: wpt-font-family${id};
        src: url(${url});
      }
      #el-${id} {
        font-family: wpt-font-family${id};
      }
    `));
    const div = document.createElement('div');
    div.setAttribute('id', 'el-' + id);
    div.appendChild(style);
    div.appendChild(document.createTextNode('x'));
    document.body.appendChild(div);

    t.add_cleanup(() => div.remove());

    return document.fonts.ready;
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
      });
  }, 'sec-fetch-site - Same origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsCrossSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsSameSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite', 'httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin', 'httpsSameSite', 'httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsCrossSite', 'httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsCrossSite', 'httpsSameSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsCrossSite', 'httpsCrossSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Cross-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin', 'httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-origin']);
      });
  }, 'sec-fetch-site - Same-Origin -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin', 'httpsSameSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsOrigin', 'httpsCrossSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Same-Origin -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsSameSite', 'httpsOrigin']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Site -> Same Origin');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsSameSite', 'httpsSameSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['same-site']);
      });
  }, 'sec-fetch-site - Same-Site -> Same-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsSameSite', 'httpsCrossSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
      });
  }, 'sec-fetch-site - Same-Site -> Cross-Site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, []))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-mode');
          assert_array_equals(headers['sec-fetch-mode'], ['cors']);
      });
  }, 'sec-fetch-mode');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, []))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_array_equals(headers['sec-fetch-dest'], ['font']);
      });
  }, 'sec-fetch-dest');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, []))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
      });
  }, 'sec-fetch-user');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsCrossSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
      });
  }, 'sec-fetch-storage-access - Cross-site');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpsSameSite']))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
      });
  }, 'sec-fetch-storage-access - Same site');
  </script>
  </body>
</html>
